+++ src/app/components/planner/PlannerActions.tsx
@@ -4,14 +4,16 @@ import LockOpenIcon from "@mui/icons-material/LockOpen";
 import SchoolIcon from "@mui/icons-material/School";
 import { Button, Card, Typography } from "@mui/joy";
 import { useSession } from "next-auth/react";
+import { useRouter } from "next/navigation";
 import { useContext } from "react";
 
 export default function PlannerActions() {
   const { data: session, status } = useSession();
+  const router = useRouter();
+  const redirectToAdminDashboard = () => router.push("planner/permissions");
 
   // TODO: remove modal context
-  const { setShowMajorsModal, setShowPermissionsModal } =
-    useContext(ModalsContext);
+  const { setShowMajorsModal } = useContext(ModalsContext);
 
   const { data: userRole } = useUserRole(session?.user.id);
   const isAdmin = userRole === "ADMIN";
@@ -29,7 +31,7 @@ export default function PlannerActions() {
     buttons.push({
       name: "Permissions",
       icon: <LockOpenIcon fontSize="large" />,
-      onClick: () => setShowPermissionsModal(true),
+      onClick: () => redirectToAdminDashboard(),
       disabled: false,
     });
   }
