+++ src/app/contexts/ModalsProvider.tsx
@@ -9,8 +9,6 @@ import useModals from "../hooks/useModals";
 export interface ModalsContextProps {
   showMajorsModal: boolean;
   setShowMajorsModal: SetShow;
-  showPermissionsModal: boolean;
-  setShowPermissionsModal: SetShow;
   showReplaceRLModal: boolean;
   setShowReplaceRLModal: SetShow;
   showMajorRequirementsEditModal: boolean;
@@ -26,8 +24,6 @@ export function ModalsProvider({ children }: { children: React.ReactNode }) {
   const {
     showMajorsModal,
     setShowMajorsModal,
-    showPermissionsModal,
-    setShowPermissionsModal,
     showReplaceRLModal,
     setShowReplaceRLModal,
     showMajorRequirementsEditModal,
@@ -43,8 +39,6 @@ export function ModalsProvider({ children }: { children: React.ReactNode }) {
       value={{
         showMajorsModal,
         setShowMajorsModal,
-        showPermissionsModal,
-        setShowPermissionsModal,
         showReplaceRLModal,
         setShowReplaceRLModal,
         showMajorRequirementsEditModal,
