+++ src/app/hooks/useModals.ts
@@ -3,7 +3,6 @@ import { useState } from "react";
 
 export default function useModals() {
   const [showMajorsModal, setShowMajorsModal] = useState(false);
-  const [showPermissionsModal, setShowPermissionsModal] = useState(false);
   const [showReplaceRLModal, setShowReplaceRLModal] = useState(false); // RL = Requirement List
   const [showMajorRequirementsEditModal, setShowMajorRequirementsEditModal] =
     useState(false);
@@ -14,8 +13,6 @@ export default function useModals() {
   return {
     showMajorsModal,
     setShowMajorsModal,
-    showPermissionsModal,
-    setShowPermissionsModal,
     showReplaceRLModal,
     setShowReplaceRLModal,
     showMajorRequirementsEditModal,
