+++ src/app/planner/permissions/page.tsx
@@ -1,0 +1,5 @@
+import Permissions from "./PermissionsPage";
+
+export default async function Page() {
+  return <Permissions />;
+}
